server/pyproject.toml
[tool.poetry]
name = "chair-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT Lisence"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.3"
google-generativeai = "^0.7.2"
dendrite-sdk = "^0.1.4"
openai = "^1.44.0"
python-dotenv = "^1.0.1"
supabase = "^2.7.4"
opencv-python = "^4.10.0.84"
flask-cors = "^5.0.0"
pyttsx3 = "^2.91"
quart = "^0.19.6"
hypercorn = "^0.17.3"
awsebcli = "^3.20.10"
quart-cors = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
